// frontend/src/modules/provider/routes/enums/route.enum.ts
export enum ProviderRouteEnum {
  ROOT = '/therapist',
  SIGN_IN = '/therapist/signin',
  SIGN_UP = '/therapist/signup',
  SIGN_UP_LAST_STEP = '/therapist/signup-last-step',
  PASSWORD_RECOVERY = '/therapist/forgot-password',
  ACCOUNT_VERIFICATION = '/therapist/verification',
  SETUP_ACCOUNT = '/therapist/setup-account',
  APPOINTMENTS = '/therapist/appointments',
  CALENDAR = '/therapist/calendar',
  PROFILE = '/therapist/profile',
  BLOCKED = '/therapist/blocked-users',
}